% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{build_game_report}
\alias{build_game_report}
\title{Game Report}
\usage{
build_game_report(gameId, model = NULL)
}
\arguments{
\item{gameId}{Game ID to build the report}

\item{model}{optional model for that season's games}
}
\value{
a list with home and away teams, their xg, and a data frame of all players and their xg for and against.
}
\description{
Get all xG for team and players for a gameId.
}
