% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{build_logistic_model}
\alias{build_logistic_model}
\title{Build xG model}
\usage{
build_logistic_model(
  season,
  num_seasons_include = 3,
  weighted = TRUE,
  save_model = TRUE
)
}
\arguments{
\item{season}{The season for which to build the model. Must be 2012 or later.}

\item{num_seasons_include}{how many seasons to include, default 3}

\item{weighted}{Whether to use weighted data (older seasons less important) or have all wights 1}

\item{save_model}{Whether to write the model to RDS file to be used later. File saved to default data directory, then \verb{.../model/[season]_model.RDS}}
}
\value{
invisibly a list of object containing the model and the metrics against the training set. The model output can be used directly with \code{predict()}
}
\description{
The model is built on a past 3 years of data as xG change over time. For example, the model for 2021 is built with 2018, 2019, 2020 seasons. 2012 and 2013 are built with one and two seasons only, respectively. Model for 2011 is built and tested on 2011 data only.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
