% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_model}
\alias{build_model}
\title{Build Model}
\usage{
build_model(season, save_model = TRUE, weighted = TRUE, model = "logistic")
}
\arguments{
\item{season}{The season for which to build the model. Must be 2011 or later.}

\item{save_model}{Whether to write the model to RDS file to be used later. File saved to default data directory, then \verb{.../model/[season]_[weighted]_[modeltype]_model.RDS}}

\item{weighted}{Whether or not to use historical weighting for logistic models. Default True}

\item{model}{Model type - logistic (using glmnet) or xgboost (using xgboost). Note that xgboost models are very slow to train}
}
\value{
a model object invisibly.
}
\description{
Build Model
}
\seealso{
build_xgb_model

build_logistic_model
}
