% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{build_player_db}
\alias{build_player_db}
\title{Build Player DB}
\usage{
build_player_db(pbp)
}
\arguments{
\item{pbp}{a pbp file, which will be the source for all of the playerID to add to db}
}
\value{
invisible player db
}
\description{
Given a pbp, build or add to playerdb all the players in that pbp
}
