% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_game_ids}
\alias{get_game_ids}
\title{getGameIDs}
\usage{
get_game_ids(season)
}
\arguments{
\item{season}{A season. Enter as a four digit number, e.g., 2018 for the 2018-2019 season}
}
\value{
a vector of gameIds in the format YYYY0XGGGG, where YYYY is the season, 0X is the game type (02 for regular season, 03 for playoffs) and GGGG is the four digit game number of the season, or the round, series and game number of playoff games
}
\description{
Returns a vector of gameIds for the season provided. Must be 2011 to present (earlier seasons do not contain play by play data).
}
