% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{get_game_xg}
\alias{get_game_xg}
\title{Get Game xG}
\usage{
get_game_xg(gameId, model = NULL)
}
\arguments{
\item{gameId}{GameID to get the xG values for}

\item{model}{Season model to use to predict xG, optional. Will load correct model for season if not provided.}
}
\value{
a named list with home_xg and away_xg
}
\description{
Get Game xG
}
\examples{
\dontrun{
get_game_xg(2011020001)
}
}
