% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_goalies}
\alias{get_goalies}
\title{Get the goalies for the provided team from the feed}
\usage{
get_goalies(feed, team)
}
\arguments{
\item{feed}{pbp feed data frame}

\item{team}{Team to get goalies for}
}
\value{
a data frame containting goalie names and player.id
}
\description{
Get the goalies for the provided team from the feed
}
