% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_player_stat}
\alias{get_player_stat}
\title{Get Player Stat}
\usage{
get_player_stat(playerId = numeric(), stat = NULL, player_db = NULL)
}
\arguments{
\item{playerId}{player ID (single or vector)}

\item{stat}{stat to query - e.g. shoots_catches, primary_position_abbreviation}

\item{player_db}{passed, or loaded from the file.}
}
\value{
a vector of \code{stat}
}
\description{
Get Player Stat
}
