% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_players_at_time}
\alias{get_players_at_time}
\title{Get all active players and goalies at any point in time}
\usage{
get_players_at_time(time, shifts)
}
\arguments{
\item{time}{Time to query, in seconds}

\item{shifts}{Shifts data frame}
}
\value{
a data.frame of players on the ice at the provided time
}
\description{
Get all active players and goalies at any point in time
}
