% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{get_xg}
\alias{get_xg}
\title{Get xG}
\usage{
get_xg(data, model_collection, metrics = FALSE)
}
\arguments{
\item{data}{The data to get an xG value rom}

\item{model_collection}{a model collection}

\item{metrics}{T/F on if you want metrics returned instead of xG values}
}
\value{

}
\description{
Get xG
}
