% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_prepped_data}
\alias{load_prepped_data}
\title{Load Prepped data}
\usage{
load_prepped_data(season)
}
\arguments{
\item{season}{Season's prepped data to load.}
}
\value{
a data frame
}
\description{
Load Prepped data
}
