% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_season_model}
\alias{load_season_model}
\title{Load Season Model}
\usage{
load_season_model(season, model = "logistic", weighted = TRUE)
}
\arguments{
\item{season}{The season's xG model to load}

\item{model}{What type of model to load - default is logistic, the default model type. also accepts xgb}

\item{weighted}{whether to load the regular or weighted model}
}
\value{
a xG model.
}
\description{
This function will load the season model from a file. If it does not exist, it will attempt to download from GitHub. Thus, if you plan to make your own xG model based on this code, please use \code{build_model[BulsinkBxG::build_model]} instead.
}
