% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_season_pbp}
\alias{load_season_pbp}
\title{Load season PBP from File}
\usage{
load_season_pbp(season)
}
\arguments{
\item{season}{Season file to load}
}
\value{
pbp of whole season
}
\description{
Load season PBP from File
}
