% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{process_game_pbp}
\alias{process_game_pbp}
\title{Process PBP}
\usage{
process_game_pbp(gameId)
}
\arguments{
\item{gameId}{the ID of the game you wish to process}
}
\value{
a pbp data frame or invisible.
}
\description{
Given a GameId, try to load and process the feed and shifts file to one pbp. Saves a file back in the default data directory named `[gameId]_pbp.RDS`
}
