% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_season}
\alias{scrape_season}
\title{Scrape season}
\usage{
scrape_season(season, skipgames = NULL, ...)
}
\arguments{
\item{season}{season 2011++ as numeric. Season as four digit year (start of season). E.g. 2018 for 20182019 season.}

\item{skipgames}{optional vector of games to skip}

\item{...}{extra parameters to pass to scrape_and_save}
}
\description{
Shortcut function to scrape a whole season's games
}
