% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_year}
\alias{scrape_year}
\title{Scrape Year}
\usage{
scrape_year(year, data_dir = "~/NHLpbp", skipgames = NULL, ...)
}
\arguments{
\item{year}{Year 2007++ as numeric}

\item{data_dir}{directory in which to save games}

\item{skipgames}{optional vector of games to skip}

\item{...}{}
}
\description{
Shortcut function to scrape a whole season's games using EW's scraper.
}
