% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{test_model}
\alias{test_model}
\title{Test a Model's performance}
\usage{
test_model(model, test_data)
}
\arguments{
\item{model}{The model to test}

\item{test_data}{The data to test agains}
}
\value{
a named list with four metrics
}
\description{
Test a Model's performance
}
