% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{compile_season}
\alias{compile_season}
\title{Compile PBP to single season file}
\usage{
compile_season(year, data_dir = "~/NHLpbp")
}
\arguments{
\item{year}{Season to compile to file}

\item{data_dir}{location of raw data}
}
\value{
pbp of whole season
}
\description{
Compile PBP to single season file
}
