% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_season_model}
\alias{load_season_model}
\title{Load Season Model}
\usage{
load_season_model(season, model = "logistic")
}
\arguments{
\item{season}{The season's xG model to load}
}
\value{
a xG model.
}
\description{
This function will load the season model from a file. If it does not exist, it will attempt to download from GitHub. Thus, if you plan to make your own xG model based on this code, please use \code{build_model[BulsinkBxG::build_model]} instead.
}
