% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_and_save}
\alias{scrape_and_save}
\title{Scrape and Save}
\usage{
scrape_and_save(
  scrapegames,
  data_dir = "~/NHLpbp",
  ignore_downloads = FALSE,
  ...
)
}
\arguments{
\item{scrapegames}{game ids to scrape}

\item{data_dir}{direcory in which to save games}

\item{ignore_downloads}{If true, will re-download games. Else only try for missing games}

\item{...}{additional parameters to pass to scraper}
}
\description{
Scrape games with EW's scraper and save the files in a directory
}
